feat: Script to convert Splunk CSV exports into JSON (#108)
Our "snapshot" data is saved in Splunk as JSON, but in plaintext lines
that also contain other data. This prevents us from exporting just the
snapshot data lines. This script allows us to export these log lines
from Splunk as a CSV file, and then running this strips the extra bits
and formats the results as proper JSON.